Обновление компонента NearPlaces
