+++ src/components/near-places/near-places.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import PropTypes from "prop-types";
 import {OfferPropType} from "../../types";
-import OfferCard from "../offer-card/offer-card";
 import {TypeCards} from "../../const";
+import NearPlaceCard from "../near-place-card/near-place-card";
 
 class NearPlaces extends React.PureComponent {
   constructor(props) {
@@ -23,20 +23,18 @@ class NearPlaces extends React.PureComponent {
   render() {
     const {offers} = this.props;
     return (
-      <React.Fragment>
-        <section className="near-places places">
-          <h2 className="near-places__title">Other places in the neighbourhood</h2>
-          <div className="near-places__list places__list">
-            {offers.map((offer) =>
-              <OfferCard
-                key={offer.id}
-                offer={offer}
-                onHover={() => this.setActiveOffer(offer)}
-                typeCard={TypeCards.NEAR_PLACES}
-              />)}
-          </div>
-        </section>
-      </React.Fragment>
+      <section className="near-places places">
+        <h2 className="near-places__title">Other places in the neighbourhood</h2>
+        <div className="near-places__list places__list">
+          {offers.map((offer) =>
+            <NearPlaceCard
+              key={offer.id}
+              offer={offer}
+              onHover={() => this.setActiveOffer(offer)}
+              typeCard={TypeCards.NEAR_PLACES}
+            />)}
+        </div>
+      </section>
     );
   }
 }
