+++ src/components/offers-list/offers-list.jsx
@@ -1,8 +1,8 @@
-import PlaceCard from "../place-card/place-card";
-import {Sort} from "../sort/sort";
 import React from "react";
 import PropTypes from "prop-types";
 import {OfferPropType} from "../../types";
+import PlaceCard from "../place-card/place-card";
+import {Sort} from "../sort/sort";
 
 export class OffersList extends React.PureComponent {
   constructor(props) {
@@ -23,21 +23,19 @@ export class OffersList extends React.PureComponent {
   render() {
     const {offers} = this.props;
     return (
-      <React.Fragment>
-        <section className="cities__places places">
-          <h2 className="visually-hidden">Places</h2>
-          <b className="places__found">{offers.length} places to stay in Amsterdam</b>
-          <Sort/>
-          <div className="cities__places-list places__list tabs__content">
-            {offers.map((offer) =>
-              <PlaceCard
-                key={offer.id}
-                offer={offer}
-                onHover={() => this.setActiveOffer(offer)}
-              />)}
-          </div>
-        </section>
-      </React.Fragment>
+      <section className="cities__places places">
+        <h2 className="visually-hidden">Places</h2>
+        <b className="places__found">{offers.length} places to stay in Amsterdam</b>
+        <Sort/>
+        <div className="cities__places-list places__list tabs__content">
+          {offers.map((offer) =>
+            <PlaceCard
+              key={offer.id}
+              offer={offer}
+              onHover={() => this.setActiveOffer(offer)}
+            />)}
+        </div>
+      </section>
     );
   }
 }
