+++ src/components/owner/owner.jsx
@@ -4,17 +4,15 @@ import {OwnerPropType} from "../../types";
 const Owner = (props) => {
   const {owner} = props;
   return (
-    <React.Fragment>
-      <div className="property__host-user user">
-        <div className={`jsx-avatar__owner property__avatar-wrapper user__avatar-wrapper ${owner.isSuper ? `property__avatar-wrapper--pro` : ``}`}>
-          <img className="property__avatar user__avatar" src={owner.avatar} width="74" height="74"
-            alt="Host avatar"/>
-        </div>
-        <span className="property__user-name">
-          {owner.name}
-        </span>
+    <div className="property__host-user user">
+      <div className={`jsx-avatar__owner property__avatar-wrapper user__avatar-wrapper ${owner.isSuper ? `property__avatar-wrapper--pro` : ``}`}>
+        <img className="property__avatar user__avatar" src={owner.avatar} width="74" height="74"
+          alt="Host avatar"/>
       </div>
-    </React.Fragment>
+      <span className="property__user-name">
+        {owner.name}
+      </span>
+    </div>
   );
 };
