+++ src/components/pages/main-page/main-page.jsx
@@ -4,6 +4,7 @@ import {OfferPropType} from "../../../types";
 import Header from "../../header/header";
 import Map from "../../map/map";
 import {OffersList} from "../../offers-list/offers-list";
+import Locations from "../../locations/locations";
 
 export class MainPage extends React.PureComponent {
   constructor(props) {
@@ -29,40 +30,7 @@ export class MainPage extends React.PureComponent {
           <main className="page__main page__main--index">
             <h1 className="visually-hidden">Cities</h1>
             <div className="tabs">
-              <section className="locations container">
-                <ul className="locations__list tabs__list">
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>Paris</span>
-                    </a>
-                  </li>
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>Cologne</span>
-                    </a>
-                  </li>
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>Brussels</span>
-                    </a>
-                  </li>
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item tabs__item--active">
-                      <span>Amsterdam</span>
-                    </a>
-                  </li>
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>Hamburg</span>
-                    </a>
-                  </li>
-                  <li className="locations__item">
-                    <a className="locations__item-link tabs__item" href="#">
-                      <span>Dusseldorf</span>
-                    </a>
-                  </li>
-                </ul>
-              </section>
+              <Locations/>
             </div>
             <div className="cities">
               <div className="cities__places-container container">
